Use an iterator for primary keys in store.Query
Introduces `objects.store.GetAllKeysWithIterator()` and replaces `indexes.FilterWithCallback()` with `indexes.FilterWithIterator()` in order to be able to iterate over the result set rather than collecting all the data. This reduces the memory usage and execution time of `store.Query`, the latter being split and transferred to the multiple calls of `Cursor.Read()`.