Add a resource field and use the up-to-date validation regex for generation
Adds a `resource` field to the TestStarname object and integrate it into the fuzzing test. Regenerated the type definitions to reflect the modified protobuf definition.
The starnames are now generated directly from the current validation regexes of the mainnet, which are fetched from https://lcd-private-iov-mainnet-2.iov.one/configuration/query/configuration