Remove object counter in objects store
The object counter was an optimization in the case we iterate over the objects with a starting offset greater than the total number of objects. It causes issues because an `crud.Store` could be created when the underlying db already has objects in it. Moreover, the optimized case is not a frequent one